+++ Changes.md
@@ -1,3 +1,4 @@
+* #1136 - Add role for report generators
 * #1135 - Allow editing of the award group order on the awards script editor page
 * #1126 - Reload scoreboard if a score is edited to be lower than it was previously
 * #1127 - Add PDF reports for top performance scores